- Optimized Obstacle generation: using Object Pooling approach now - Optimized LoopingBackground object - Optimized data types - Optimized initial values (cached where possible) - Optimized EyeMover - Optimized Player object - Optimized music, sounds and sprites - Removed unused tags - Removed unused code

+++ Assets/Scripts/ButtonLanguage.cs
@@ -5,7 +5,7 @@ using UnityEngine.UI;
 
 public class ButtonLanguage : MonoBehaviour
 {
-    public int localeId;
+    public ushort localeId;
 
     private void Start()
     {
@@ -32,7 +32,7 @@ public class ButtonLanguage : MonoBehaviour
         GetComponent<Image>().color = tempColor;
     }
 
-    public void ChangeLocaleProxy(int localeId)
+    public void ChangeLocaleProxy(ushort localeId)
     {
         FindObjectOfType<SettingsManager>().ChangeLocale(localeId);
     }
